Filtering methods for the session user

+++ buylist/views/buylist_views.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from ..forms import BuyListForm
 from ..models import BuyList
@@ -7,8 +8,7 @@ from ..services import buylist_service
 
 @login_required()
 def buylist_list(request):
-    buylists = buylist_service.buylist_list()
-
+    buylists = buylist_service.buylist_list(request.user)
     return render(request, 'buylist/buylist_list.html', {'buylists': buylists})
 
 
@@ -20,7 +20,7 @@ def register_buylist(request):
         if form_buylist.is_valid():
             name = form_buylist.cleaned_data['name']
 
-            new_buylist = BuyList(name=name)
+            new_buylist = BuyList(name=name, user=request.user)
             buylist_service.register_buylist(new_buylist)
 
             return redirect('buylist_list_route')
@@ -35,6 +35,10 @@ def register_buylist(request):
 @login_required()
 def edit_buylist(request, id):
     buylist_db = buylist_service.buylist_list_id(id)
+
+    if buylist_db.user != request.user:
+        return HttpResponse('Não Permitido!')
+
     form_buylist = BuyListForm(request.POST or None, instance=buylist_db)
 
     if form_buylist.is_valid():
@@ -52,6 +56,9 @@ def edit_buylist(request, id):
 @login_required()
 def remove_buylist(request, id):
     buylist_db = buylist_service.buylist_list_id(id)
+
+    if buylist_db.user != request.user:
+        return HttpResponse('Não Permitido!')
 
     if request.method == 'POST':
         buylist_service.remove_buylist(buylist_db)
